Move window_split() to windows.py
