+++ rasterio/windows.py
@@ -785,3 +785,40 @@ class Window:
         Window
         """
         return intersection([self, other])
+
+
+def window_split(height, width, max_pixels=1048576):
+    """Divide the row-column domain of a dataset into smaller windows.
+
+    This function determines the window size such that windows have no
+    more than max_pixels number of pixels. Windows are roughly square,
+    except at the right and bottom edges of the domain, and have integer
+    offsets and lengths.
+
+    Parameters
+    ----------
+    height : int
+        Domain height.
+    width : int
+        Domain width.
+    max_pixels : int, optional
+        The maximum number of pixels per window.
+
+    Returns
+    -------
+    list of Windows
+    """
+    length = min(int(math.floor(math.sqrt(max_pixels))), height, width)
+    ncols = int(math.ceil(width / length))
+    nrows = int(math.ceil(height / length))
+    chunk_windows = []
+
+    for col in range(ncols):
+        col_offset = col * length
+        w = min(length, width - col_offset)
+        for row in range(nrows):
+            row_offset = row * length
+            h = min(length, height - row_offset)
+            chunk_windows.append(((row, col), Window(col_offset, row_offset, w, h)))
+
+    return chunk_windows
